refactor: change fps type from f32 to f64 for FFmpeg consistency

+++ src/core/stream_info.rs
@@ -68,7 +68,7 @@ pub enum StreamInfo {
 
         /// The frames per second (FPS) of the video stream, represented as a floating point number.
         /// It is calculated from the `avg_framerate` field (avg_framerate.num / avg_framerate.den).
-        fps: f32,
+        fps: f64,
 
         /// The rotation of the video stream in degrees. This value is retrieved from the metadata.
         /// Common values are 0, 90, 180, and 270.
@@ -284,7 +284,7 @@ pub fn find_video_stream_info(url: &str) -> Result<Option<StreamInfo>> {
         let bit_rate = (*codec_parameters).bit_rate;
         let pixel_format = (*codec_parameters).format;
         let video_delay = (*codec_parameters).video_delay;
-        let fps = avg_frame_rate.num as f32 / avg_frame_rate.den as f32;
+        let fps = avg_frame_rate.num as f64 / avg_frame_rate.den as f64;
 
         // Fetch the rotation info from metadata (if present)
         let rotate = metadata
@@ -729,7 +729,7 @@ pub fn find_all_stream_infos(url: &str) -> Result<Vec<StreamInfo>> {
                     let video_delay = (*codec_parameters).video_delay;
                     let r_frame_rate = (*stream).r_frame_rate;
                     let sample_aspect_ratio = (*stream).sample_aspect_ratio;
-                    let fps = avg_frame_rate.num as f32 / avg_frame_rate.den as f32;
+                    let fps = avg_frame_rate.num as f64 / avg_frame_rate.den as f64;
 
                     // Fetch the rotation info from metadata (if present)
                     let rotate = metadata
